Add docker-compose-plugin to packages to be installed
The versioning of the docker compose plugin is different than that of
the docker-compose Python package.  The assessors require a version
>=2.2.3, which forces the installation of docker-compose-plugin
package.

Also update the Molecule test code in accordance with these changes,
and refactor the code to be more understandable and more in line with
the way we have been writing such tests lately.